Reformat endpoints
